refactor: segunda parte da arquitetura limpa
Novos diretorios com base no que foi debatido:
- adapter/driver/ -> presenters/
- adapter/driven -> datasources/
- interfaces -> interfaceAdapters
