+++ src/datasources/database/models/categoriaModel.ts
