+++ src/datasources/database/models/produtoImagensModel.ts
