+++ src/datasources/database/seeders/cria-categorias.ts
