+++ src/interfaceAdapters/controllers/categoriaController.ts
