+++ src/presenters/api/routers/schemas/categoriaRouter.schema.ts
